Check dispatcher tasks

+++ Dev/AppUnits/Atdi.AppUnits.Sdrn.DeviceServer.Repositories/SaveMessages.cs
@@ -30,12 +30,14 @@ namespace Atdi.AppUnits.Sdrn.DeviceServer.Repositories
             {
                 var fileName = MakeFileName(typeof(T).Name.ToString());
                 var fullPath = Path.Combine(this._messageFolder, fileName);
-
-                IFormatter formatter = new BinaryFormatter();
-                using (MemoryStream stream = new MemoryStream())
+                lock (fileName)
                 {
-                    formatter.Serialize(stream, receivedMessage);
-                    File.WriteAllBytes(fullPath, stream.ToArray());
+                    IFormatter formatter = new BinaryFormatter();
+                    using (MemoryStream stream = new MemoryStream())
+                    {
+                        formatter.Serialize(stream, receivedMessage);
+                        File.WriteAllBytes(fullPath, stream.ToArray());
+                    }
                 }
                 return fullPath;
             }
@@ -57,7 +59,10 @@ namespace Atdi.AppUnits.Sdrn.DeviceServer.Repositories
             {
                 if (File.Exists(fileName))
                 {
-                    File.Delete(fileName);
+                    lock (fileName)
+                    {
+                        File.Delete(fileName);
+                    }
                     return true;
                 }
                 else
@@ -81,13 +86,18 @@ namespace Atdi.AppUnits.Sdrn.DeviceServer.Repositories
 
             try
             {
-                var fullPath = Path.Combine(this._messageFolder, fileName);
-
-                IFormatter formatter = new BinaryFormatter();
-                using (MemoryStream stream = new MemoryStream())
+                lock (fileName)
                 {
-                    formatter.Serialize(stream, receivedMessage);
-                    File.WriteAllBytes(fullPath, stream.ToArray());
+                    var fullPath = Path.Combine(this._messageFolder, fileName);
+                    if (File.Exists(fullPath))
+                    {
+                        IFormatter formatter = new BinaryFormatter();
+                        using (MemoryStream stream = new MemoryStream())
+                        {
+                            formatter.Serialize(stream, receivedMessage);
+                            File.WriteAllBytes(fullPath, stream.ToArray());
+                        }
+                    }
                 }
                 return true;
             }
